Backend/requirement.txt
mysqlclient
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv